Prevent building theme previews in production postinstall script

+++ install.sh
@@ -17,11 +17,13 @@ else
 	(cd client/overlay; webpack -d --watch) &
 fi
 
-echo "Building theme previews..."
-THEMES_DIR=./themes
-THEMES_OUTPUT_DIR=${THEMES_ROOT:-./data/themes}
-rm -rf $THEMES_OUTPUT_DIR/*
-mkdir -p $THEMES_OUTPUT_DIR
-for theme in $(ls $THEMES_DIR); do
-	./workers/theme/bundle $THEMES_DIR/$theme $THEMES_OUTPUT_DIR/$theme
-done
+if [ "$NODE_ENV" != "production" ]; then
+	echo "Building theme previews..."
+	THEMES_DIR=./themes
+	THEMES_OUTPUT_DIR=${THEMES_ROOT:-./data/themes}
+	rm -rf $THEMES_OUTPUT_DIR/*
+	mkdir -p $THEMES_OUTPUT_DIR
+	for theme in $(ls $THEMES_DIR); do
+		./workers/theme/bundle $THEMES_DIR/$theme $THEMES_OUTPUT_DIR/$theme
+	done
+fi
